fix(api,ui): Fix autoscaling target parsing (#380)
* Allow setting of autoscaling targets on the ui without constraints on number of decimal places

* Add parsing of autoscaling targets to int for cpu and rps

* Update min and step size for autoscaling target field

* Update unit tests and error messages

+++ ui/src/services/version/RouterVersion.js
@@ -95,7 +95,7 @@ export class RouterVersion {
         resource_request: this.resource_request,
         autoscaling_policy: {
           ...this.autoscaling_policy,
-          target: parseInt(this.autoscaling_policy.target),
+          target: parseFloat(this.autoscaling_policy.target),
         },
         protocol: this.protocol,
       },
@@ -112,7 +112,7 @@ export class RouterVersion {
             resource_request: this.enricher.resource_request,
             autoscaling_policy: {
               ...this.enricher.autoscaling_policy,
-              target: parseInt(this.enricher.autoscaling_policy.target),
+              target: parseFloat(this.enricher.autoscaling_policy.target),
             },
           }
           : {
@@ -136,7 +136,7 @@ export class RouterVersion {
                     ...this.ensembler.docker_config,
                     autoscaling_policy: {
                       ...this.ensembler.docker_config.autoscaling_policy,
-                      target: parseInt(
+                      target: parseFloat(
                         this.ensembler.docker_config.autoscaling_policy.target
                       ),
                     },
@@ -148,7 +148,7 @@ export class RouterVersion {
                       ...this.ensembler.pyfunc_config,
                       autoscaling_policy: {
                         ...this.ensembler.pyfunc_config.autoscaling_policy,
-                        target: parseInt(
+                        target: parseFloat(
                           this.ensembler.pyfunc_config.autoscaling_policy.target
                         ),
                       },
